Spike with dtspec 0.7.1
Use dtspec 0.7.1 in order to make use of jinja support in rendering dtspec specifications.

It is definitely very beneficial to make use of template values like:

- {{ TODAY }}
- {{ TOMORROW }}
- {{ dbt_source('dbt_shop', 'raw_products') }}

There appears though to exist an issue with the current dtspec version when working with
multiple spec files. By default, they all get compiled to one huge `compiled_spec.yml`
which can lead to clashes between the identifier / source /target definitions.

Also in case of not defining a global factory the "compilation" of the spec fails
because by default it is expected to have at least one global factory (minItems = 1).